refactor(playground-ui): move self-contained tool badges into playground-ui
ToolApprovalButtons, AskUserBadge, AskUserTool and CodeModeBadge only
depend on the tool-call context and design-system components that
already live in playground-ui, so they can be shared with hosts other
than Studio. Their tests move with them; the badges/types.ts re-export
is dropped in favour of importing the ask-user types directly.
