+++ packages/playground/src/lib/ai-ui/tools/__tests__/agent-badge-routing-decision.test.tsx
@@ -9,7 +9,7 @@ vi.mock('@mastra/playground-ui/domains/chat/components/network-choice-metadata-d
   NetworkChoiceMetadataDialogTrigger: mockNetworkChoiceMetadataDialogTrigger,
 }));
 
-vi.mock('../badges/tool-approval-buttons', () => ({
+vi.mock('@mastra/playground-ui/domains/chat/tools/badges/tool-approval-buttons', () => ({
   ToolApprovalButtons: mockToolApprovalButtons,
 }));
