+++ packages/playground/src/lib/ai-ui/tools/badges/agent-badge-wrapper.tsx
@@ -1,10 +1,10 @@
 import { toAISdkV5Messages } from '@mastra/ai-sdk/ui';
 import type { MessageMetadata } from '@mastra/playground-ui/domains/chat';
 import { LoadingBadge } from '@mastra/playground-ui/domains/chat/components/loading-badge';
+import type { ToolApprovalButtonsProps } from '@mastra/playground-ui/domains/chat/tools/badges/tool-approval-buttons';
 import type { AgentMessage } from './agent-badge';
 import { AgentBadge } from './agent-badge';
 import { resolveToChildMessages } from './resolve-child-messages';
-import type { ToolApprovalButtonsProps } from './tool-approval-buttons';
 import { useAgentMessages } from '@/hooks/use-agent-messages';
 
 interface SubAgentToolResult {
