--- packages/playground/src/lib/ai-ui/tools/badges/types.ts
@@ -1,6 +1,0 @@
-export type {
-  AskUserOption,
-  AskUserPayload as AskUserSuspendPayload,
-  AskUserResult,
-  AskUserSelectionMode,
-} from '@mastra/playground-ui/components/ai/ask-user';
